LUN-3: Move Interfaces from menuItem.ts
