--- src/interfaces/menuItem.ts
@@ -1,9 +1,0 @@
-import { ReactElement } from 'react'
-
-export interface IMenuItemProps {
-  selected: boolean
-  onPress: () => void
-  icon: string
-  title: string
-  children: ReactElement
-}
